+++ packages/signer/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## 0.21.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fuel-ts/hasher@0.21.1
+  - @fuel-ts/address@0.21.1
+  - @fuel-ts/keystore@0.21.1
+  - @fuel-ts/math@0.21.1
+
+## 0.21.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fuel-ts/hasher@0.21.0
+  - @fuel-ts/address@0.21.0
+  - @fuel-ts/keystore@0.21.0
+  - @fuel-ts/math@0.21.0
+
 ## 0.20.0
 
 ### Patch Changes